Reworked hook module so that platform-dependent code is split up in a submodule
